requirements.txt
spotipy~=2.23.0
python-dotenv~=1.0.1
tinytag~=1.10.1
Flask~=3.0.3
openai~=1.13.3
requests~=2.31.0
Levenshtein~=0.25.0
fuzzywuzzy~=0.18.0